move tests to mod
